requirements.txt
greenlet==3.1.1
pyodbc==5.2.0
python-decouple==3.8
SQLAlchemy==2.0.36
typing_extensions==4.12.2
azure-identity==1.19.0